requirement.txt
gradio
python-dotenv
langchain_core
langsmith
langchain_openai
langchain
langchain-deepseek
pandas
langchain_community
chromadb
dashscope

openpyxl